# Conflicts:
#	MediaEngine_Windows/MediaEngine_Windows/MediaEngine_Windows.vcxproj.filters
#	OthersLib/JNI/Android.mk
#	VideoEngineController/AudioCallHeader.cpp
#	VideoEngineController/AudioCallHeader.h
#	VideoEngineController/AudioCallSession.cpp
#	VideoEngineController/AudioCallSession.h
#	VideoEngineController/AudioLiveHeader.cpp
#	VideoEngineController/AudioLiveHeader.h
#	VideoEngineController/EncodedFramePacketizer.cpp

+++ VideoEngineController/SendingThread.h
@@ -32,7 +32,7 @@ public:
 private:
 	int GetSleepTime();
 
-	int m_nTimeStampOfChunck;
+	long long m_nTimeStampOfChunck;
 	int m_nTimeStampOfChunckSend;
 
 	CVideoCallSession* m_pVideoCallSession;
